commit code

+++ src/Component/HandelAdmin.js
@@ -1,12 +1,18 @@
 import api from "../apis";
+import HandelUpdate from "./HandelUpdate";
 
 const HandelAdmin = async () => {
   const productList = document.querySelector("#productList");
 
   const deleProduct = (id) => {
-    console.log(id);
-    api.delete(`products/${id}`);
-    window.location.reload();
+    try {
+      if (confirm("Bạn có muốn xoá sản phẩm")) {
+        api.delete(`products/${id}`);
+        window.location.reload();
+      }
+    } catch (error) {
+      console.log(error);
+    }
   };
 
   productList.addEventListener("click", (e) => {
@@ -15,21 +21,38 @@ const HandelAdmin = async () => {
       const productId = target.dataset.id;
       deleProduct(productId);
     }
+    if (target.classList.contains("btn-edit")) {
+      const productId = target.dataset.id;
+      HandelUpdate(productId);
+    }
   });
+
   // hien thi du lieu
   const { data } = await api.get(`/products`);
   if (data) {
+    data.forEach((item) => {
+      if (typeof item.gallery === "string") {
+        item.gallery = [item.gallery];
+      }
+    });
     const html = data.map(
-      (item) => `
+      (item) => /*html*/ `
         <tr>
         <td>${item.id}</td>
         <td>${item.title}</td>
         <td>${item.price}</td>
-        <td><img src="${item.gallery[0]}" alt="Ảnh sản phẩm"></td>
+       <td><img src="${
+         item.gallery && item.gallery.length > 0 ? item.gallery[0] : ""
+       }" alt="Ảnh sản phẩm"></td>
+
         <td>${item.description}</td>
         <td class="action">
-       <a href="edit/${item.id}"> <button class="btn-edit" data-id="${item.id}">Edit</button></a>
-        <button class="btn-dele" data-id="${item.id}">Delete</button>
+       <a href="/edit/${item.id}"> <button class="btn-edit" data-id="${
+        item.id
+      }"><i class="bi bi-pencil-square"></i></button></a>
+        <button class="btn-dele" data-id="${
+          item.id
+        }"><i class="bi bi-trash3"></i></button>
         </td>
         </tr>
         `
